dags/requirements.txt
dbt-core==1.1.1
dbt-snowflake
apache-airflow-providers-snowflake==1.3.0